requirements.txt
celery>=5.4.0
django>=5.1.7
django-celery-beat>=2.7.0
djangorestframework>=3.15.2
pandas>=2.2.3
sqlalchemy>=2.0.39
